requirements.txt
streamlit
tensorflow==2.13.0
tensorflow-hub==0.16.1
numpy<1.24.0
pillow